Rename the service
